nGaussCurveFitter working, reporting update in progress
