+++ src/main/java/utils/RadialProfiler.java
@@ -40,17 +40,7 @@ import java.util.*;
 
 public class RadialProfiler {
 
-    public SortedMap<Double, Double> oCorrMap;
-    public SortedMap<Double, Double> sCorrMap;
-
-    public Gaussian gaussian;
-
-    //3 Parameter double: Normalization, Mean, Sigma
-    public double[] gaussFitParameters;
-
-    public Double confidence;
-
-    public Double rSquared;
+    protected CorrelationData correlationData;
 
     private long[] dimensions;
 
@@ -76,23 +66,23 @@ public class RadialProfiler {
     }
 
     public void calculateOCorrProfile(RandomAccessibleInterval origCorrelation) {
-        oCorrMap = new TreeMap<>();
+        correlationData.oCorrMap = new TreeMap<>();
 
-        calculateSingleProfile(origCorrelation, oCorrMap);
+        calculateSingleProfile(origCorrelation, correlationData.oCorrMap);
     }
 
     public void calculateSCorrProfile(RandomAccessibleInterval subtractedCorrelation) {
-        sCorrMap = new TreeMap<>();
+        correlationData.sCorrMap = new TreeMap<>();
 
-        calculateSingleProfile(subtractedCorrelation, sCorrMap);
+        calculateSingleProfile(subtractedCorrelation, correlationData.sCorrMap);
     }
 
     public void calculateBothProfiles(RandomAccessibleInterval origCorrelation, RandomAccessibleInterval subtractedCorrelation) {
-        oCorrMap = new TreeMap<>();
-        sCorrMap = new TreeMap<>();
+        correlationData.oCorrMap = new TreeMap<>();
+        correlationData.sCorrMap = new TreeMap<>();
 
-        calculateSingleProfile(origCorrelation, oCorrMap);
-        calculateSingleProfile(subtractedCorrelation, sCorrMap);
+        calculateSingleProfile(origCorrelation, correlationData.oCorrMap);
+        calculateSingleProfile(subtractedCorrelation, correlationData.sCorrMap);
     }
 
     private <T extends RealType> void calculateSingleProfile(RandomAccessibleInterval<T> input, Map<Double, Double> output) {
@@ -136,137 +126,5 @@ public class RadialProfiler {
         tempMap.forEach((key,value) -> {
             output.put(key.doubleValue(), (value[0]/value[1]));
         });
-    }
-
-    public void fitGaussianCurve(){
-        gaussFitParameters = CurveFit(sCorrMap);
-
-        gaussian = new Gaussian(gaussFitParameters[0], Math.abs(gaussFitParameters[1]), gaussFitParameters[2]);
-
-        if(oCorrMap != null)
-            confidence = (areaUnderCurve(sCorrMap, gaussFitParameters[1], gaussFitParameters[2]) / areaUnderCurve(oCorrMap, gaussFitParameters[1], gaussFitParameters[2]));
-
-        rSquared = getRsquared();
-    }
-
-    private double[] CurveFit(SortedMap<Double, Double> inputMap) {
-        double maxLoc = 0;
-        double max = 0;
-        WeightedObservedPoints obs = new WeightedObservedPoints();
-
-        /*First need to determine the maximum value in order to set the weights for the fitting, and determine its
-         * location for instances where the mean is close to zero (in order to mirror the data, this has to be done
-         * for a good fit)
-         */
-        for (Double d : inputMap.keySet()) {
-            if (inputMap.get(d) > max) {
-                maxLoc = d;
-                max = inputMap.get(d);
-            }
-        }
-
-        /* this next loop adds values below zero that mirror values equidistant from the opposite side of the peak value (max at maxLoc).
-         * This is done for fits where the means are near zero, as this data is zero-bounded. Not mirroring the data results
-         * in very poor fits for such values. We can't simply mirror across 0 as this will create a double-peak
-         * for any data where the peak is near but not at zero.
-         * It would be preferable to fit the data using a truncated gaussian fitter, but I could not find any available
-         * java class that performs such a fit and my own attempts were unsuccessful.
-         */
-        if(maxLoc == inputMap.firstKey()){
-            maxLoc = 0.0;
-        }
-        double finalMaxLoc = maxLoc;
-
-        inputMap.forEach((key,value) -> {
-            obs.add(key, value);
-            if (key > 2 * finalMaxLoc) {
-                obs.add(((2 * finalMaxLoc) - key), value);
-            }
-        });
-
-        double [] output = null;
-        nGaussianCurveFitter curveFitter = nGaussianCurveFitter.create();
-        try{
-            System.out.println("Initializing Curvefitter");
-            //todo: simple test, change to nGaussianCurveFitter and put in curveCount of 2, then print output
-            output = curveFitter.withCount(2).withMaxIterations(100).fit(obs.toList());
-        }
-        catch(TooManyIterationsException ignored){}
-
-        /*
-         * Have to check if the curve was fit to a single noise spike, something that came up quite a bit during initial testing.
-         * If not fit to a noise spike, values are returned with no further processing, if it is, the data is averaged
-         * with nearest neighbors and another fit is attempted. This usually only needs a single round of averaging.
-         *
-         * We can use the pixel scale to test for this, as the SD of the spatial correlation should never be less than
-         * the pixel size.
-         */
-        if (output == null || output[2] <= scale[0] || output[1] < -scale[0] || output[0] < 0) {
-            for (Double windowSize = scale[0]/10; (output == null || output[2] <= scale[0] || output[1] < 0) && windowSize <= (scale[0]/2); windowSize += scale[0]/10){
-                obs.clear();
-                SortedMap<Double,Double> averaged = MovingAverage.averagedMap(inputMap, windowSize);
-                max = 0;
-                maxLoc = 0;
-                for (Double d : averaged.keySet()) {
-                    if (averaged.get(d) > max) {
-                        maxLoc = d;
-                        max = averaged.get(d);
-                    }
-                }
-                if(maxLoc == averaged.firstKey()){
-                    maxLoc = 0.0;
-                }
-                double finalMaxLoc1 = maxLoc;
-
-                averaged.forEach((key, value) -> {
-                    obs.add(key, value);
-                    if (key > 2 * finalMaxLoc1) {
-                        obs.add(((2 * finalMaxLoc1) - key), value);
-                    }
-                });
-                try {
-                    output = curveFitter.withCount(2).withMaxIterations(100).fit(obs.toList());
-                }
-                catch(TooManyIterationsException ignored){}
-
-            }
-        }
-        for (int i = 0; i < output.length; i++) {
-            System.out.println(output[i]);
-        }
-        if(output == null|| output[2] <= scale[0] || output[1] < -scale[0] || output[0] < 0){
-            gaussFitParameters = new double[]{0, inputMap.lastKey(), inputMap.lastKey()};
-            confidence = -1.0;
-            rSquared = -1.0;
-            gaussian = new Gaussian(1, -1,  1);
-
-            throw new NullPointerException("Could not fit Gaussian curve to data");
-        }
-        return output;
-    }
-
-    private double areaUnderCurve(Map<Double, Double> map, double mean, double sigma) {
-
-        double auc = 0;
-
-        for (Double d : map.keySet()) {
-            if ((mean - (3 * sigma)) < d && d < (mean + (3 * sigma))) {
-                auc += map.get(d);
-            }
-        }
-
-        return auc;
-    }
-
-    private double getRsquared(){
-        final double[] residualsSum = {0};
-        final double[] totalSum = {0};
-        final double rangeMean = sCorrMap.subMap(gaussFitParameters[1] - (3 * gaussFitParameters[2]), gaussFitParameters[1] + (3 * gaussFitParameters[2])).values().stream().mapToDouble(Double::doubleValue).average().getAsDouble();
-
-        sCorrMap.subMap(gaussFitParameters[1] - (3 * gaussFitParameters[2]), gaussFitParameters[1] + (3 * gaussFitParameters[2])).forEach((key, value) -> {
-            residualsSum[0] += Math.pow(value - gaussian.value(key), 2);
-            totalSum[0] += Math.pow(value - rangeMean, 2);
-        });
-        return (1-(residualsSum[0]/totalSum[0]));
     }
 }
